+++ ui/src/main/java/com/jin/ui/order/OrderScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.order.ui
+package com.jin.ui.order
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
@@ -16,9 +16,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.ui.cart.CartScreen
-import com.jin.honey.feature.order.ui.content.OrderHistoryScreen
+import com.jin.ui.order.content.OrderHistoryScreen
 import com.jin.feature.ui.state.DbState
 import com.jin.feature.ui.state.UiState
